refactor(titiler-app): split main.py into focused modules and improve logging
Extract config, logging setup, FastAPI dependencies, and request middleware
  into separate files. Add explicit Python logging configuration driven by
  TTL_LOG_LEVEL, request/response timing middleware, and more targeted log
  calls throughout the config resolution path.
